requirements.txt
torch
pytorch-lightning
awkward
autopep8
boost-histogram
docopt
hydra-core
jupyter
jupyter-book
matplotlib
mplhep
notebook
numpy
papermill
pre-commit
scikit-learn
scikit-optimize
scipy
seaborn
setGPU
tensorboard
tqdm
uproot
vector
wheel
zenodo-get